Got sidetracked on thinking about DB Admin tools for the framework.
Made interfaces that give a jumping off point. As DB admin is unique for each RDBMS you have to create different classes for each.

Anyway, don't think I need this for the guide project but all the same.

Also, thinking I need to fork off the Framework so it is running a different git line.  Time for it.